dataflow/gpu-workers/requirements.txt
Pillow
rasterio

# apache-beam[gcp] -- already installed when running in Dataflow.
# tensorflow -- installed in the Dockerfile with conda for GPU support.
